+++ backend/backend/handlers/indexing/sqsBucketSync.py
@@ -28,7 +28,7 @@ lambda_client = boto3.client('lambda')
 dynamodb_client = boto3.client('dynamodb')
 logger = safeLogger(service_name="sqsBucketSync")
 
-reservedPrefixFolders = ['temp-upload', 'temp-uploads', 'preview','previews', 'pipeline', 'piplines']
+reservedPrefixFolders = ['temp-upload', 'temp-uploads', 'preview','previews', 'pipeline', 'pipelines']
 
 # Environment variables
 try: